pyproject.toml
[project]
name = "trustehragent"
version = "0.1.0"
requires-python = "==3.10.*"
dependencies = [
    "aiofiles==24.1.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asyncer==0.0.8",
    "attrs==25.3.0",
    "autogen-agentchat==0.2.40",
    "azure-ai-inference==1.0.0b9",
    "azure-ai-projects==1.0.0b10",
    "azure-common==1.1.28",
    "azure-core==1.34.0",
    "azure-identity==1.21.0",
    "azure-search-documents==11.5.2",
    "backports-tarfile==1.2.0",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "contourpy==1.3.2",
    "cryptography==44.0.2",
    "cycler==0.12.1",
    "deprecated==1.2.18",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "docker==7.1.0",
    "exceptiongroup==1.2.2",
    "flaml==2.3.4",
    "fonttools==4.57.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "isodate==0.7.2",
    "jaraco-context==6.0.1",
    "jiter==0.9.0",
    "jsonlines==3.1.0",
    "jsonref==1.1.0",
    "kiwisolver==1.4.8",
    "levenshtein==0.23.0",
    "matplotlib==3.7.1",
    "more-itertools==10.7.0",
    "msal==1.32.3",
    "msal-extensions==1.3.1",
    "openai==1.76.2",
    "opentelemetry-api==1.32.1",
    "packaging==25.0",
    "pandas==2.1.4",
    "pillow==11.2.1",
    "protobuf==5.29.4",
    "pycparser==2.22",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pyjwt==2.10.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-levenshtein==0.23.0",
    "pytz==2025.2",
    "rapidfuzz==3.13.0",
    "regex==2024.11.6",
    "requests==2.32.3",
    "seaborn==0.13.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "termcolor==2.4.0",
    "tiktoken==0.9.0",
    "tqdm==4.67.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "wolframalpha==5.0.0",
    "wrapt==1.17.2",
    "xmltodict==0.14.2",
    "zipp==3.21.0",
]
